services/gemzy-moments-api/server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.24.0
supabase>=2.5.0
stripe>=9.0.0
python-dotenv>=1.0.0
google-cloud-storage>=2.18.0
httpx>=0.27.0
email-validator>=2.1.1
